+++ AudioPlaybackConnector2/include/app/ApplicationHost.hpp
@@ -5,6 +5,7 @@
 #include <app/SettingsWindowPresenter.hpp>
 #include <app/SingleInstanceGuard.hpp>
 
+#include <services/CommandLineControlServer.hpp>
 #include <services/NotificationService.hpp>
 #include <services/SettingsController.hpp>
 #include <services/TrayController.hpp>
@@ -44,6 +45,7 @@ private:
     void InitializeTray();
     void InitializeNotifications();
     void InitializeDeviceManager();
+    void InitializeCommandLineControl();
     void SetupDeviceEvents();
     void TeardownDeviceEvents();
     void TryAutoReconnect();
@@ -62,6 +64,7 @@ private:
 
     void ShowSettingsWindow();
     void ExitApplication();
+    apc::control::Response HandleControlCommand(apc::control::Request const& request);
     void PerformTeardown(bool saveLastConnected) noexcept;
     void RunOnUIThread(std::function<void()> work);
 
@@ -96,6 +99,7 @@ private:
 
     std::shared_ptr<NotificationService> m_notificationService;
     std::shared_ptr<TrayController> m_trayController;
+    CommandLineControlServer m_commandLineControlServer;
     DeviceEventRouter m_deviceEventRouter;
     SingleInstanceGuard m_singleInstanceGuard;
     static inline UINT s_wmTaskbarCreated = 0;